# my-app/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "Shopping Service API Specialized in coffee",
    "version": "1.0",
    "title": "bean cafe API",
    "termsOfService": "about:blank",
    "contact": {
      "name": "Hongrok Lim",
      "url": "https://github.com/rokong/",
      "email": "hongrr123@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host": "localhost:8080",
  "basePath": "/bean-cafe",
...
  "paths": {
...
    "/review": {
      "get": {
        "tags": [
          "Review"
        ],
        "summary": "get reviews",
        "description": "[get reviews (productId), (optionCd), (userNm)]",
        "operationId": "getReviewsUsingGET",
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "review",
            "description": "review",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Review"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Review"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/review/": {
      "post": {
        "tags": [
          "Review"
        ],
        "summary": "create review",
        "description": "[userNm, orderId, productId, optionCd, rate, content, (isVisible)]",
        "operationId": "createReviewUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "review",
            "description": "review",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Review"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Review"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/review/{id}": {
      "get": {
        "tags": [
          "Review"
        ],
        "summary": "get review",
        "description": "get review",
        "operationId": "getReviewUsingGET",
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Review"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "Review"
        ],
        "summary": "delete review",
        "operationId": "removeReviewUsingDELETE",
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      },
      "patch": {
        "tags": [
          "Review"
        ],
        "summary": "update review",
        "description": "[(rate), (content), (isVisible)]",
        "operationId": "updateReviewUsingPATCH",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "review",
            "description": "review",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Review"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Review"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
...
    "Review": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32",
          "example": 1,
          "description": "review id"
        },
        "userNm": {
          "type": "string",
          "example": "customer1",
          "description": "writer name"
        },
        "orderId": {
          "type": "integer",
          "format": "int32",
          "example": 2,
          "description": "order id"
        },
        "productId": {
          "type": "integer",
          "format": "int32",
          "example": 3,
          "description": "product id"
        },
        "optionCd": {
          "type": "string",
          "example": "0101",
          "description": "option code"
        },
        "rate": {
          "type": "number",
          "format": "double",
          "example": 7.2,
          "description": "rate"
        },
        "content": {
          "type": "string",
          "example": "not bad.",
          "description": "content"
        },
        "isVisible": {
          "type": "boolean",
          "example": true,
          "description": "is visible"
        },
        "updateDt": {
          "type": "string",
          "example": "2020-12-31",
          "description": "update date"
        }
      },
      "title": "Review",
      "description": "Review from purchased products"
    },
...
  }
}